+++ src/mutex/MemcachedMutex.php
@@ -24,18 +24,18 @@ class MemcachedMutex extends SpinlockMutex
      *
      * @param string    $name     The lock name.
      * @param Memcached $memcache The connected Memcached API.
-     * @param int       $timeout  The time in seconds a lock expires, default is 3.
+     * @param float     $timeout  The time in seconds a lock expires, default is 3.
      *
      * @throws \LengthException The timeout must be greater than 0.
      */
-    public function __construct(string $name, Memcached $memcache, int $timeout = 3)
+    public function __construct(string $name, Memcached $memcache, float $timeout = 3)
     {
         parent::__construct($name, $timeout);
 
         $this->memcache = $memcache;
     }
 
-    protected function acquire(string $key, int $expire): bool
+    protected function acquire(string $key, float $expire): bool
     {
         return $this->memcache->add($key, true, $expire);
     }
